+++ pat/src/pat_ring_buffer.h
@@ -2,7 +2,7 @@
 #include <stdint.h>
 
 typedef struct PATRingBuffer {
-    uint8_t* data;
+    uint32_t* data;
     size_t read_position;
     size_t write_position;
     size_t size;
@@ -11,8 +11,8 @@ typedef struct PATRingBuffer {
 
 PATRingBuffer* pat_ring_buffer_create(size_t capacity);
 
-size_t pat_ring_buffer_read(PATRingBuffer *ring_buffer, uint8_t *data, size_t data_size);
+size_t pat_ring_buffer_read(PATRingBuffer *ring_buffer, uint32_t *data, size_t data_size);
 
-size_t pat_ring_buffer_write(PATRingBuffer *ring_buffer, uint8_t *data, size_t data_size);
+size_t pat_ring_buffer_write(PATRingBuffer *ring_buffer, uint32_t *data, size_t data_size);
 
 void pat_ring_buffer_free(PATRingBuffer *ring_buffer);
